requirements.txt
Flask==2.0.1
Flasgger
torch==2.0.0
torchvision==0.15.0
Pillow==8.3.2
python-dotenv==0.19.0
pymongo==3.12.0
Flask
gunicorn
werkzeug
pygame==2.5.0
google-generativeai==0.3.0
gTTS==2.3.2
